code cleanup, fix log writes
